+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- On macOS, fix `set_cursor_visible` hides cursor outside of window.
 - On macOS, fix `CursorEntered` and `CursorLeft` events fired at old window size.
 - On macOS, fix error when `set_fullscreen` is called during fullscreen transition.
 - On all platforms except mobile and WASM, implement `Window::set_minimized`.